refactor: 에러 메시지만 받는 NotFoundAnimalException 생성자를 생성한다.
