+++ src/test/java/com/clova/anifriends/domain/animal/service/AnimalServiceTest.java
@@ -1,7 +1,8 @@
 package com.clova.anifriends.domain.animal.service;
 
 import static com.clova.anifriends.domain.animal.support.fixture.AnimalFixture.animal;
-import static com.clova.anifriends.domain.shelter.support.fixture.ShelterFixture.shelter;
+import static com.clova.anifriends.domain.shelter.support.ShelterFixture.shelter;
+import static com.clova.anifriends.domain.shelter.support.ShelterImageFixture.shelterImage;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchException;
 import static org.mockito.ArgumentMatchers.anyLong;
@@ -39,6 +40,7 @@ class AnimalServiceTest {
         void findAnimalByIdByVolunteer() {
             // given
             Shelter shelter = shelter();
+            shelter.setShelterImage(shelterImage(shelter));
             Animal animal = animal(shelter);
             FindAnimalByVolunteerResponse expected = FindAnimalByVolunteerResponse.from(animal);
