+++ src/test/java/com/clova/anifriends/domain/shelter/support/ShelterFixture.java
@@ -8,7 +8,7 @@ public class ShelterFixture {
     private static final String SHELTER_PASSWORD = "shelterPassword";
     private static final String SHELTER_ADDRESS = "shelterAddress";
     private static final String SHELTER_ADDRESS_DETAIL = "shelterAddressDetail";
-    private static final String SHELTER_NAME = "보호소이름";
+    private static final String SHELTER_NAME = "shelterName";
     private static final String PHONE_NUMBER = "010-1234-5678";
     private static final String SPARE_PHONE_NUMBER = "010-8765-4321";
     private static final boolean IS_OPENED_ADDRESS = true;
