--- src/test/java/com/clova/anifriends/domain/shelter/support/fixture/ShelterFixture.java
@@ -1,33 +1,0 @@
-package com.clova.anifriends.domain.shelter.support.fixture;
-
-import static com.clova.anifriends.domain.shelter.support.ShelterImageFixture.shelterImage;
-
-import com.clova.anifriends.domain.shelter.Shelter;
-
-public class ShelterFixture {
-
-    private static final String SHELTER_EMAIL = "shelterEmail@email.com";
-    private static final String SHELTER_PASSWORD = "shelterPassword";
-    private static final String SHELTER_ADDRESS = "shelterAddress";
-    private static final String SHELTER_ADDRESS_DETAIL = "shelterAddressDetail";
-    private static final String SHELTER_NAME = "shelterName";
-    private static final String PHONE_NUMBER = "010-1234-5678";
-    private static final String SPARE_PHONE_NUMBER = "010-8765-4321";
-    private static final boolean IS_OPENED_ADDRESS = true;
-
-    public static Shelter shelter() {
-        Shelter shelter = new Shelter(
-            SHELTER_EMAIL,
-            SHELTER_PASSWORD,
-            SHELTER_ADDRESS,
-            SHELTER_ADDRESS_DETAIL,
-            SHELTER_NAME,
-            PHONE_NUMBER,
-            SPARE_PHONE_NUMBER,
-            IS_OPENED_ADDRESS
-        );
-        shelter.setShelterImage(shelterImage(shelter));
-
-        return shelter;
-    }
-}
